refactor: SystemTime 객체를 core 로 이동
