+++ backend/src/main/java/com/morak/back/team/application/TeamService.java
@@ -1,6 +1,6 @@
 package com.morak.back.team.application;
 
-import com.morak.back.appointment.domain.SystemTime;
+import com.morak.back.core.domain.SystemTime;
 import com.morak.back.auth.domain.Member;
 import com.morak.back.auth.domain.MemberRepository;
 import com.morak.back.auth.exception.MemberNotFoundException;
