+++ backend/src/main/java/com/morak/back/team/domain/ExpiredTime.java
@@ -1,6 +1,6 @@
 package com.morak.back.team.domain;
 
-import com.morak.back.appointment.domain.SystemTime;
+import com.morak.back.core.domain.SystemTime;
 import com.morak.back.core.exception.CustomErrorCode;
 import com.morak.back.team.exception.TeamDomainLogicException;
 import java.time.LocalDateTime;
