+++ backend/src/test/java/com/morak/back/appointment/AppointmentCreateRequestFixture.java
@@ -1,6 +1,6 @@
 package com.morak.back.appointment;
 
-import com.morak.back.appointment.domain.SystemTime;
+import com.morak.back.core.domain.SystemTime;
 import com.morak.back.appointment.ui.dto.AppointmentCreateRequest;
 import com.morak.back.appointment.ui.dto.AvailableTimeRequest;
 import com.morak.back.core.domain.FakeSystemTime;
