+++ backend/src/test/java/com/morak/back/appointment/acceptance/AppointmentAcceptanceTest.java
@@ -31,7 +31,7 @@ import static com.morak.back.appointment.AppointmentCreateRequestFixture.총_진
 import static org.assertj.core.api.Assertions.assertThat;
 
 import com.morak.back.AcceptanceTest;
-import com.morak.back.appointment.domain.SystemTime;
+import com.morak.back.core.domain.SystemTime;
 import com.morak.back.appointment.ui.dto.AppointmentAllResponse;
 import com.morak.back.appointment.ui.dto.AppointmentCreateRequest;
 import com.morak.back.appointment.ui.dto.AppointmentResponse;
