+++ backend/src/test/java/com/morak/back/appointment/application/AppointmentSchedulerTest.java
@@ -4,7 +4,7 @@ import static org.assertj.core.api.Assertions.assertThat;
 
 import com.morak.back.appointment.domain.Appointment;
 import com.morak.back.appointment.domain.AppointmentRepository;
-import com.morak.back.appointment.domain.SystemTime;
+import com.morak.back.core.domain.SystemTime;
 import com.morak.back.core.domain.Code;
 import com.morak.back.core.domain.menu.MenuStatus;
 import com.morak.back.support.ServiceTest;
